Fix everything for now + norm

+++ ft_bzero.c
@@ -6,13 +6,15 @@
 /*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 13:58:24 by lilefebv          #+#    #+#             */
-/*   Updated: 2024/11/05 14:01:34 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 11:47:13 by lilefebv         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *pointer, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *pointer, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
