+++ ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 15:10:03 by lilefebv          #+#    #+#             */
-/*   Updated: 2024/11/06 17:48:02 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 11:44:48 by lilefebv         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int	ft_strlcat(char *dest, const char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	dest_len;
-	unsigned int	src_len;
+	size_t	i;
+	size_t	dest_len;
+	size_t	src_len;
 
 	src_len = ft_strlen(src);
 	dest_len = ft_strlen(dest);
